build harmonizer v0.x and harmonizer v2.x in the same workspace (#8)
building v0 and v2 next to each other gives us the benefit of being able to see both packages at the same time, not have to track changes across multiple branches, and allows us to share types with our helper crates more easily.

xtask will be in charge of renaming harmonizer-0 orharmonizer-2 to just harmonizer depending on which version it is publishing and then immediately changing it back for development in this repository. the two versions must never be published at the same time as the packages are not named properly.