chore: creates supergraph CLI
the goal here is to have a `supergraph` binary that works very similarly for both fed1 and fed2 so that it's easy to shell out to it from rover and be able to deserialize the json it returns. this pr achieves that! i think. it's looped into the "stage" just like harmonizer is (though it won't ever be published to crates.io) and its versions are kept up to date.

there's only one helper crate now with two features, build and config, so you can include what you want when you want, and composition will always return a `apollo_federation_types::build::BuildResult` that should remain stable for... pretty much forever? i've tried to make it so that if we make breaking changes to the serialization of the type that older versions of the parser should be able to catch any new fields with the `other: crate::UncaughtJson` escape hatch, but this should be tested.

**the thinking here is that from Rover, we can download the _latest_ `supergraph` binary that is compatible with your operating system and the "platform version" you specify in `supergraph.yml`. since the API is all based around the same `BuildResult` type, Rover can just import it and attempt to deserialize the output from `supergraph` into it (it should always succeed!)**

this will let folks have a similar experience with Rover that they do with Apollo Studio. a lockfile would pair nicely with this, if you run compose with `--locked` it will check the lock file for the `supergraph` binary version to download and it will use that one to run composition.