router-bridge: Introduce native `graphql` (JS) `validate` to the `plan`
While Federation's `buildQueryPlan` function does a fair bit of validation
in order to properly build our plan (and leverages `graphql`'s [`validate`]
function do most of that, it *does not* do all of the [default validations]
that are specified.

To be thorough, we must call `validate`, as we did [prior] to updating to the
new Federation v2 alpha query planner.

This also introduces a number of tests which exercise specific validations
which we know as of today are not part of federation:

 - invalid_graphql_validation_1_is_caught: NoFragmentCyclesRule
 - invalid_graphql_validation_2_is_caught: ScalarLeafsRule
 - invalid_graphql_validation_3_is_caught: NoUnusedFragmentsRule

I chose three random but different types to make sure we're capturing it.

[prior]: https://github.com/apollographql/federation/commit/9cb220135d776043acfe0cee411277a8f2ce0976
[`validate`]: https://github.com/graphql/graphql-js/blob/95dac43fd4bff037e06adaa7cfb44f497bca94a7/src/validation/validate.ts#L38
[default validations]: https://github.com/graphql/graphql-js/blob/95dac43fd4bff037e06adaa7cfb44f497bca94a7/src/validation/specifiedRules.ts#L76-L103

Closes: https://github.com/apollographql/federation-rs/issues/25