chore: Make harmonizer snapshots
Fixes #44

deno_core advocates for us to use snapshots instead of loading scripts and running them.
One of the recent side effects is that deno_core prints `rerun-if-changed` directives when loading scripts manually.

This commit moves the scripts loading mechanism to the build (in `build_harmonizer.rs`) so the messages mentionned above won't display at runtime.