fix: generate deno snapshots in build.rs
In #58 I added a snapshots generation phase to our harmonizer packages.

However I created a bug because the snapshot generation phase would happen when the harmonizer-0 and harmonizer-2 packages are generated.

This means downstream users who would rely on published harmonizer crates would get snapshots generated by an other computer / architecture.

This triggered a bug that would fail any harmonizer call when relying on the published crate.

This commit adds a build.rs script to the published crates, so that snapshots are created when downstream binaries are built.