fix: QueryPlanner race
This commit fixes a Race that would happen when racing queries against the planner.

It bumps dependencies and introduces a Channel multiplexer. Long story short, Each Request made against deno gets assigned an Uuid, which allows us to send the response to the right requester.

During this refactoring, we've also seen an issue that would happen when the TS runtime would try to return javascript Errors.

Since an Error is not iterable, it would not pass the Deno -> Rust fence because we rely on serialization. We thus catch errros and turn them into an { message, name, stack } structure, that we send to the deno side.

Moreover this commit reworks the Query Planning instanciation and allows us to return clearer error messages.