Always generate query plan even when pretty plan serialization fails
We recently added the ability for the Router to have access to the "pretty
printed" query plan in https://github.com/apollographql/federation-rs/pull/154.

The `prettyFormatQueryPlan` function can throw an error (internally) when
its unable to print the plan, however, no failure to produce a pretty
printed query plan should be severe enough stop execution of an otherwise
valid plan.  The pretty printed aspect is purely for diagnostics purposes
and for consumption by humans.