try out an idea for protecting the query planner from abuse
The idea is to maintain a HashSet of queries where we couldn't notify
our requester that query planning has completed.

They become, effectively, blacklisted and the next time we receive that
request, we just reject it.

This would help a lot with the kind of problem TV2 encountered.

The are some drawbacks with the approach though. There are many
legitimate resons for a requester to disappear, so it seems harsh to
always blacklist (see comments in code for ways to mitigate).

Anyway, I think it's worth exploring further and as a side effect I got
rid of Uuid and rely on the rust hasher for ID generation.