improve idea by removing blacklist
and adding a cache for undelivered plans.

This removes the complexity of having to decide when to blacklist a plan
and avoids repeated plan calculation by holding the plan in a cache.

Subsequent requests for the same query will be satisfied by the
undelivered plan cache (which will then find itself in the router query
plan cache).