Fix deserialization of error codes
The code for error coming from composition through harmonizer were not
correctly handled, because when deserializing the JSON, the `code` was
excepted at the top-level of the error object, while in practice it
is within an `extensions` sub-object.

This seems like an oversight seems the error type itself has the `code`
field with a sub-structure of the `extensions` field, and so it should
deserialize correctly but wasn't because of a "rogue"
`#[serde(flatten)]`. This commit thus simply remove that incorrect
flatening.