Install Volta w/ `cargo install` to overcome lack of Linux ARM binaries (#203)
This aims to resolve a blocker encountered when trying to
land https://github.com/apollographql/federation-rs/pull/202, which
appears to be a new transient failure in `volta` and is preventing us
from releasing Federation 2.1.4 in both `harmonizer` and
`router-bridge`. The error encountered there is:

```
Releases for non x64 architectures are not currently supported
```

Which is produced by the install script
[here](https://github.com/volta-cli/volta/blob/main/dev/unix/volta-install.sh#L129-L136).

Since the Linux ARM binaries are not published and the install script
seems to struggle with the request to install from source we need to
just build in our more traditional ways using `cargo install`.

We may want to consider pinning the `volta` version if this continues to
be problematic, but this should unblock us in the near term. (I did not
bother bisecting if there was a previous version on which this worked.)