chore: add is_config to BuildErrors (#205)
added an `is_config` field to `BuildErrors` to provide more helpful
suggestions in Rover when a `supergraph.yaml` cannot be resolved.
Assumption here is that you will always have either config errors _or_
composition errors. It's unfortunate that we didn't specify those types
quite properly when this was first launched as we're stuck w/the
existing JSON structure for this kind of error (meaning we should have
had `CompositionErrors` and `ConfigErrors` instead of just
`BuildErrors`, but this is where we are).