updateable Planner (#259)
Make the planner reusable by supporting schema changes in an existing JS
runtime. Since the router might still be serving requests while we are
creating a new router pipeline with a new schema, we must support
multiple concurrent planners in the same runtime.

Here is how it works:
- when sending a schema to the JS side, we pass along a random schema
id, created with the rust side `Planner` instance
- on every request from this `Planner`, the schema id will be sent
- when sending a new schema, we create a new `Planner` instance with the
same `JsWorker` but a different schema id
- on the JS side, we have a map `schema id -> planner` so we know which
planner to use for the request
- when a `Planner` is dropped, it still sends an `Exit` message, but
with its schema id. On the JS side, we remove the corresponding planner
from the map. If the map is empty, we stop the task