router-bridge: add option to disable validation + marker property for validation errors (#355)
Part of https://github.com/apollographql/router/pull/3134

This makes GraphQL validation optional so the router can disable it if
it's doing validation in Rust already.

To support running both validations for a while, errors from validation
are marked with `.validationError = true`. The router can use that to
tell the difference between validation errors and other errors and
compare the result with Rust validation. We can remove that later when
we are confident in the Rust validation implementation.

When adding the `validationError` property to errors from
`api_schema.ts`, I would lose them during serialization between JS and
Rust, and I couldn't really figure that out. But the `api_schema`
functions are only used during tests in the router as far as I can see,
in real operation the schema validation errors also come from the query
planner. So it doesn't really matter if we dont have this in the api
schema code.

To do:
- [x] I think it'll be more robust to mark validation errors, instead of
marking every other type of error...