Fix returning thrown errors from QP
Fixes https://github.com/apollographql/router/issues/4783

The Rust side is expecting a single string in `exception.stacktrace`,
but JS was providing an array of strings. It doesn't really make sense
to split `error.toString()` as normally the first line will be the error
message, and only the later lines will be part of the stack trace, so
I think it's best if we use a single string.