ci: fix CMake install by explicitly using `cmake.install` on windows (#462)
I can't explain why, but `choco install cmake -y` appears to no longer
work, as the `cmake.install` package can't be found. I tried pinning to
an older version on a previous branch, and in this branch I tried to
update chocolatey itself as well, but it keeps failing. I thought it
might be a networking configuration issue, but it turns out targeting
`cmake.install` directly instead of through the metapackage works.

I'd propose just using `cmake.install` at least for now. As a side
benefit, there is a command line flag to automatically configure the
PATH, so we don't need to do it separately.