Update deno_core to 0.280.0 (latest) in federation-2/harmonizer.
This is a big upgrade, and we should probably apply a similar upgrade in
router-bridge (though it will be trickier there), but the harmonizer's
usage of deno_core is limited/constrained enough that I think we can be
confident this upgrade is safe as long as tests pass.

More immediately, updating deno_core will finally fix the
deno_core::serde_v8 deserialization bug I described in my comment
https://github.com/apollographql/federation-rs/pull/479#issuecomment-2110342100,
which has been mysteriously manifesting only on CircleCI, and only when
building with `cargo build --release --target aarch64-apple-darwin`.

I verified this update fixes the problem by SSHing into the CircleCI
build machine and building the supergraph binary there, and I will
re-verify it once the CircleCI checks for this PR complete.