Avoid deleting all tags before pushing tags.
I don't understand why the original author of this code thought it was
important to delete all local tags, but that process now takes a
considerable amount of time and is unnecessary. Instead, we can just
push specific tags instead of using the `git push --tags` command.

The `git pull` command requires a remote tracking branch to be set.
Since this is not a guaranteed condition, it's better to skip the
unnecessary `git pull` command and assume we're already up-to-date
before the tag_release function is called.