Remove Fed 1 from main (#491)
This simplifies the repo into a single Cargo workspace which removes a
bunch of extra infrastructure. The idea is we can have a `v1` or
`support/v1` branch where we can still make tweaks to fed 1 (e.g.,
security fixes) without it being in the way for new fed 2 work.