Move some `apollo-composition` types into `apollo-federation-types` (#611)
The current dependency tree ends up requiring that Rover pull in
`apollo-federation` even though it never needs that code. This has
resulted in a bunch of maintenance burden in trying to pin the correct
combination of dependencies.

This PR moves all the types that `apollo-language-server` depends on
into `apollo-federation-types` (which Rover already includes anyway).
The result is that `apollo-language-server` no longer requires
`apollo-composition`, so neither does Rover, and therefore neither pull
in `apollo-federation`.

I also added to the changelog that we should be pinning
`apollo-federation` as a reminder that when we do our next release we
should use `=2.0.0-preview.4` to avoid confusing churn in the packages
that _do_ still depend on `apollo-composition`.

<!-- CNN-527 -->